\name{ccaPP-package}
\alias{ccaPP-package}
\alias{ccaPP}
\docType{package}
\title{
\packageTitle{ccaPP}
}
\description{
\packageDescription{ccaPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccaPP}
\packageIndices{ccaPP}
}
\author{
\packageAuthor{ccaPP}

Maintainer: \packageMaintainer{ccaPP}
}
\keyword{package}
