% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxCor.R
\name{maxCorProj}
\alias{maxCorProj}
\title{(Robust) maximum correlation via projections through the data points}
\usage{
maxCorProj(x, y, method = c("spearman", "kendall", "quadrant", "M",
  "pearson"), control = list(...), standardize = TRUE,
  useL1Median = TRUE, fallback = FALSE, ...)
}
\arguments{
\item{x, y}{each can be a numeric vector, matrix or data frame.}

\item{method}{a character string specifying the correlation functional to
maximize.  Possible values are \code{"spearman"} for the Spearman
correlation, \code{"kendall"} for the Kendall correlation, \code{"quadrant"}
for the quadrant correlation, \code{"M"} for the correlation based on a
bivariate M-estimator of location and scatter with a Huber loss function, or
\code{"pearson"} for the classical Pearson correlation (see
\code{\link{corFunctions}}).}

\item{control}{a list of additional arguments to be passed to the specified
correlation functional.  If supplied, this takes precedence over additional
arguments supplied via the \code{\dots} argument.}

\item{standardize}{a logical indicating whether the data should be
(robustly) standardized.}

\item{useL1Median}{a logical indicating whether the \eqn{L_{1}}{L1} medians
should be used as the centers of the data sets in standardization (defaults
to \code{TRUE}).  If \code{FALSE}, the columnwise centers are used instead
(columnwise means if \code{method} is \code{"pearson"} and columnwise
medians otherwise).}

\item{fallback}{logical indicating whether a fallback mode for robust
standardization should be used.  If a correlation functional other than the
Pearson correlation is maximized, the first attempt for standardizing the
data is via median and MAD.  In the fallback mode, variables whose MADs are
zero (e.g., dummy variables) are standardized via mean and standard
deviation.  Note that if the Pearson correlation is maximized,
standardization is always done via mean and standard deviation.}

\item{\dots}{additional arguments to be passed to the specified correlation
functional.}
}
\value{
An object of class \code{"maxCor"} with the following components:
\item{cor}{a numeric giving the maximum correlation estimate.}
\item{a}{numeric; the weighting vector for \code{x}.}
\item{b}{numeric; the weighting vector for \code{y}.}
\item{centerX}{a numeric vector giving the center estimates used in
standardization of \code{x}.}
\item{centerY}{a numeric vector giving the center estimates used in
standardization of \code{y}.}
\item{scaleX}{a numeric vector giving the scale estimates used in
standardization of \code{x}.}
\item{scaleY}{a numeric vector giving the scale estimates used in
standardization of \code{y}.}
\item{call}{the matched function call.}
}
\description{
Compute the maximum correlation between two data sets via projection pursuit
based on projections through the data points, with a focus on robust and
nonparametric methods.
}
\details{
First the candidate projection directions are defined for each data set
from the respective center through each data point.  Then the algorithm
scans all \eqn{n^2} possible combinations for the maximum correlation,
where \eqn{n} is the number of observations.
}
\examples{
data("diabetes")
x <- diabetes$x
y <- diabetes$y

## Spearman correlation
maxCorProj(x, y, method = "spearman")
maxCorProj(x, y, method = "spearman", consistent = TRUE)

## Pearson correlation
maxCorProj(x, y, method = "pearson")

}
\seealso{
\code{\link{maxCorGrid}}, \code{\link{ccaProj}},
\code{\link{corFunctions}},
}
\author{
Andreas Alfons
}
\keyword{multivariate}
\keyword{robust}
