\name{ccaPP-package}
\alias{ccaPP-package}
\alias{ccaPP}
\docType{package}
\title{
\packageTitle{ccaPP}
}
\description{
\packageDescription{ccaPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccaPP}
\packageIndices{ccaPP}
}
\author{
\packageAuthor{ccaPP}

Maintainer: \packageMaintainer{ccaPP}
}
\references{
A. Alfons, C. Croux and P. Filzmoser (2016) Robust maximum association between 
data sets: The \R Package \pkg{ccaPP}.  \emph{Austrian Journal of Statistics}, 
\bold{45}(1), 71--79.

A. Alfons, C. Croux and P. Filzmoser (2016) Robust maximum association 
estimators.  \emph{Journal of the American Statistical Association}.  DOI 
10.1080/01621459.2016.1148609.  In press.
}
\keyword{package}
